:- missing LUSDst.sol fix from previous commit (note: testing looks good now) 	- deployed v0.4 		# tLUSDst_0.4: 0x014D7caE54F7fDd22eBac48C049F64271b84c8b4 NOTE: - so far testing with this versions looks good, results… 	- sending PLS to contract register account balance correctly 	- checking collective balances as KEEPER displays correctly 	- executing payOutBST to buy and burn and mint, indeed works correctly 	- re-checking collective balances as KEEPER displays correctly after payOutBSt 	- KEEPER_maintenace works correctly just fine
NEXT:
	- integrate test web dapp GUI … using template: trinity.html

#========================================================#
# LUSDst
#========================================================#
TODO:
	- integrate ‘lock’ modifier for public functions as needed

#========================================================#
#========================================================#
# LEGACY …
#========================================================#
:- NEXT …
	- cleanup the __main__ execution in _lpcleaner.py
		needs to be decoupled and more modular 
		should also create a single dependency for common functions
			(between _lpcleaner.py & _keeper.py)
	- design & integrate algo’s 1 & 2
		we now have all token reserves and ‘getAmountsIn’ values

#========================================================#
NEXT:
	- flash-swap (_lpcleaner.py & UniswapFlashQuery.sol
		integrate algo’s 1 & 2 from whiteboard design
		note: all token data already retrieved and organized

	- meme coin sniping
		searching newly launched pairs (high trade count, high liquidity, recently active)
		waiting on waves, using MCAP on dexscreener

	- flash-loans with arb
		by comparing ratios of UniswapV2Pairs from on-chain (like in flashbots simple-arbitrage)
		 instead of by last sale price from off-chain dexscreener (like in arb_searcher.py)

	- flash-loans: w/ PHIAT
		$8k = $32k TVL @ 8% APY = ~$49/wk earned

	N/A - flash-loans: w/ Atropa (check for puzzles in green)
		cover buy side of mint puzzles - NOT ENOUGH LIQUIDITY
	N/A - flash-loans: w/ LL & PHIAT
		deposit <-> borrow loop - NOT PROFITABLE
	N/A - flash-loans: w/ PulseX & INC
		farm reward snipe attempt - NOT POSSIBLE


	- arb-bot fishing: automate TBF deploy -> LP creation
	- reddit: openAI image generator bot
	- reddit: $100 contract deployments
	- ollama AI: AWS hosting w/ stable diffusion
		ref: HOW-TO_ollama_ai.txt

#========================================#

#=================================#
setting…
	set min cashout -> set / checked in ‘SET_USER_WITHDRAW_REQUESTED’
	set max tw per day -> set / checked in ‘UPDATE_USER_SHILL_APPR_EARNS’ -> ‘usr_shill_limit_reached’
	set default rates -> set / check default pay rates in ‘ADD_NEW_TG_USER’
	set tw conf exp -> set / check in ‘ADD_USER_SHILL_TW’ -> ‘valid_tg_user_tw_conf’ -> ‘exp_tw_conf’

steps… for current LP testing
	- deploy contract
	- deposit $6 in PLS to the contract
	- pay self $5 in BST through the contract
	- create $10 LP (5:5) -> BST:WPLS
	- turn on dex options (true true false)
	- pay self $1 in BST
	- review payout & collective balances logs

#=================================#
PYTHON trinity bot updates
	WORKING - cleanup all TG cmd return copy outputs

PYTHON finish web3 dapp
	- finished REST endpoint testing from javascript to python
	- deploy nginx and uwsgi on aws server
	- register domain for aws server
	- integrate ssl cert with ‘how-to_ssl_certbot_letsencrypty.txt’

NEXT:
	- fix / finish _event_listener.py (utilize event hash technique from ‘write_with_hash’)
	- solve verify & publish issue on block explorers
	- mailforwarding (NY -> PP, CA -> CB)

NEXT:
	- update copy and verbage on depth’s latest work in ‘bearshare’ repo
	- db design to store question and log correct answers
		requirements need to be finalized for how payments need to
		 occur and be logged, in comparison with how ‘shills’ table works
	- db procs needed for this design
		to get latest question
		to check answer against latest question
		log correct answer by user
		 	and update user_earns table for correct answer
	- python:
		finish db calls within ‘oracle_answer’ & ‘oracle_question’
	- integrate openAI asking a new question once an hour or so
		regardless if previous question was answered yet or not
	- db design and procs on ‘ai_context’
		ie. storing and getting AI_ROLE & BOT_INFO
	- update _bst_keeper.py to support using contract ABI
		w/ full function signatures, instead of just func_hashes
#========================================================#
#========================================================#

+++ src/_abi.py
@@ -254,7 +254,7 @@ LUSDst_FUNC_MAP_READ = {
     # read functions (bst legacy)
     "#------------BST-legacy------------#": ["xxxxxxxx", [], []],
     "KEEPER()": ["862a179e", [], ['address']],
-    "KEEPER_collectiveStableBalances(bool,uint256)": ["cf0c8683", ['bool','uint256'], ['uint64','uint64','uint256','int64']],
+    "KEEPER_collectiveStableBalances(bool,uint256)": ["cf0c8683", ['bool','uint256'], ['uint64','uint64','int64','uint256']],
     # "KEEPER_getRatios(uint256)": ["ffa21500", ['uint256'], ['uint32','uint32']],
     
     "ACCT_USD_BALANCES(address)": ["c67483dc", ["address"], ['uint64']],
